feat(users): change type of id => BigInteger

+++ src/main/java/chen/binflare/dao/UserCredentialDAO.java
@@ -2,5 +2,7 @@ package chen.binflare.dao;
 
 import chen.binflare.entity.UserCredentialDO;
 
-public interface UserCredentialDAO extends BaseRepository<UserCredentialDO, Long> {
+import java.math.BigInteger;
+
+public interface UserCredentialDAO extends BaseRepository<UserCredentialDO, BigInteger> {
 }
